refactor(persistence): split audit methods into separate file (#309)
Move record_settings_audit, record_settings_audit_on, and
list_settings_audit out of repository/mod.rs into a sibling
repository/audit.rs module. The methods remain on Repository via
a separate `impl Repository` block in the new file, so no callers
change.

Brings mod.rs from 1088 lines to 986 lines, under the 1000-line
hard-fail cap enforced by scripts/dev/quality-check.sh.
